ON-15817: Fix flex array declaration macro in zf_emu
The __DECLARE_FLEX_ARRAY macro which we now use in onload, results in a
build failure for zf_emu as gcc will complain that a flex array is not
the last element in a struct. Additionally, since zf_emu is technically
C++ and not C an empty struct has a non-zero size. This commit addresses
both of these issues by redefining the __DECLARE_FLEX_ARRAY macro in
zf_emu so that the flex array is now just zero-length and removes the
empty struct declaration.