ON-15724: ensure Jenkins-only tests have proper attributes
Adding the testjenkins target meant that some tests were not given
the correct environments, and would therefore always fail. In this
change, zfstackdump is moved to its original test group as it was
not observed to fail. The remaining test, zftimestamping, came
from the testloop target, so the testjenkins target now takes
the environment from testloop to run this test.