ON-15875: continue with build in zf_mkdist_build if CC is not set
This script would try its best to ensure that CC was set to a good
value, but would default to RHEL-specific defaults if no other method
was used to provide hints. This meant that the script would fallover
on non-RHEL OSes even if compilation could work without setting CC.
As such, this change allows CC to have no value, and will just try
to continue with compilation as it might work in most cases anyway.