ON-15255: use ef_vi packet in timestamp calls
Commit da134ec0d456 (ON-15255: use new unified ef_vi timestamping
API, 2024-07-01) was an improper patch, and used to wrong packet
reference; instead of referencing the ef_vi rxpkt it instead tried
to reference the TCPDirect packet store. This change references
the correct packet to correctly extract a timestamp.

This change also teaches zf_emu to expect sub-nanosecond resolution
timestamps, rather than the previous clobbering of the lower-order
bits in the nanoseconds field.