ON-15875: add --name option to zf_mkdist
This option is passed through to zf_make_tarball and is used as the
start of the generated tarball name. Previously, zf_mkdist would
specify this as "zf", but this was dropped in Commit 05f83369e204
("ON-15875: use onload tarball in Jenkins pipeline") which meant
the new default was "tcpdirect" to match the existing Jenkins
pipeline. However, this broke other consumers of zf_mkdist that
expected a tarball starting with "zf" to be generated. As such,
this change both adds the option for this to be configured, to be
used in the Jenkins pipeline, and reverts to the previous default
name prefix, "zf".