ON-15895: handle TX completions for in-flight retransmitted packets
If we have not yet processed a TX completion for a given packet and
the RTO timer expires, we will retransmit it via PIO or by copying
it to a new buffer. When adding TX timestamping, these packets would
be marked to be freed when the TX completion was processed. However,
we would still prepare a TX timestamp report for it, thus expecting
the report to be completed.

The problem with this is that the flag used on retransmit to request
the packet gets freed immediately, `ZF_REQ_ID_PROTO_TCP_FREE`, would
indeed free the packet immediately without informing the TX report
system of this TX completion event. The next sent packet would then
populate the TX report for this retransmission with the timestamp in
its TX completion event. Subsequent packets would also experience
this off-by-one failure.

This failure mode can be triggered by sending some data, then not
performing the reactor until the RTO timer must have expired, at
which point performing the reactor again would handle the timer
expiration before checking for the probably-already-present TX
completion, meaning we get this specific flavour of retransmit.

To fix this, the send flag has been updated so that we process the
TX completion, to be consistent with other retransmissions which
get TX reports. This also required a minor update to correctly
label this report as a retransmission when preparing the report.