ON-15945: Add option to build RPM from system installed onload
Previously, in order to build the tcpdirect binary rpm you would have to
define the `onload_tarball` macro so that the build system would have
all the requisite libraries and headers. Now that we have the
`onload-devel` package (and have included citools.a in `onload`), we can
use the system installed onload when building tcpdirect and do away with
the macro.

The functionality of `onload_tarball` is still preserved in case a user
wants to build from a specific version of onload but doesn't want to
install it on their system. Additionally, if you want to build the
socket shim using `zf_mkdist` you must specify the onload tarball to use
as the shim requies `ZF_DEVEL=1` which in turn requires a full copy of
the onload source code.