ON-16261: replace versions.yaml with versions.env
This patch simplifies the format used to describe version information
and changes all consumers of the versions.yaml file to use the new
key-value pair versions.env file. This also fixes an issue where
generated packages would report the TCPDirect version as the list of
all version in the old versions.yaml file.