ON-16844: avoid double-free of tcp segments
If a segment that has already been freed is freed again, this can lead
to a crash where we end up freeing more packets than have been
allocated. There are several checks throughout the code to avoid this,
especially in the form of assertions, but this is behaviour that
should be guarded against in release builds. This patch adds tracking
to each tcp_seg to ensure it is only freed once.