ON-16844: avoid superfluous iterations in tcp_free_segs
This function operates on the stack's packet pool, which is indexed
by a uint16_t, however this function would cast these indices to
unsigned integers which is guaranteed to be at least 16-bits wide,
but is typically 32-bits wide. As a result, when the end index was
less than the start index, such as when begin=65535 and end=1, the
loop would process 2^32 - begin iterations before wrapping around
instead of processing 2^16 - begin iterations before wrapping.
This patch explicitly uses 16-bit unsigned integers to avoid this.